support PrivateMethodPatch on go1.17, go1.18
